完成教师添加练习的前后端逻辑 题目分数修改为int

+++ app/src/main/java/com/edu_app/model/Question.java
@@ -19,7 +19,7 @@ public class Question implements Serializable {
     private String questionType;
     private String question;
     private String answer;
-    private double score;
+    private int score;
 
     public Question() {}
 
@@ -75,11 +75,11 @@ public class Question implements Serializable {
         this.answer = answer;
     }
 
-    public double getScore() {
+    public int getScore() {
         return score;
     }
 
-    public void setScore(double score) {
+    public void setScore(int score) {
         this.score = score;
     }
 }
